代金券活动

+++ main4-spread/main4-spread-api/src/main/java/com/shigu/main4/spread/service/VoucherService.java
@@ -16,7 +16,7 @@ public interface VoucherService {
 
     String flag = "voucher";
 
-    void obtainVoucher(Long userId,Long uploadNum,Long pemId);
+    void obtainVoucher(Long userId,Integer uploadNum,Long pemId);
 
     List<VoucherVO> selVoucherList(Long userId,List<Long> pemIds);
